refactor(game): move player increment logic to reducer
