+++ src/game/services/Rules/index.test.ts
@@ -13,8 +13,6 @@ import { IGame, IPlayer } from '../../types'
 import { updatePlayer } from '../../reducers/update-player'
 import { RandomNumber } from '../../../services/RandomNumber'
 
-import { updateGame } from '../../reducers/update-game/index'
-
 import { Rules } from '.'
 
 const player1 = stubPlayer()
@@ -126,23 +124,12 @@ describe('Rules', () => {
   })
 
   describe('processTurnEnd', () => {
-    test('changes from player 1 to player 2', () => {
+    test('increments player', () => {
       jest.spyOn(RandomNumber, 'generate').mockReturnValueOnce(1)
       const game = Rules.processGameStart([player1, player2])
       const [player1Id] = Object.keys(game.table.players)
 
       const newGame = Rules.processTurnEnd(game)
-
-      expect(newGame.currentPlayerId).toEqual(player1Id)
-    })
-
-    test('changes from player 2 to player 1', () => {
-      jest.spyOn(RandomNumber, 'generate').mockReturnValueOnce(1)
-      const game = Rules.processGameStart([player1, player2])
-      const [player1Id, player2Id] = Object.keys(game.table.players)
-
-      let newGame = updateGame(game, { currentPlayerId: player2Id })
-      newGame = Rules.processTurnEnd(game)
 
       expect(newGame.currentPlayerId).toEqual(player1Id)
     })
